Addressed orderOfOperations function
I created a test to RUN every fixture and found a bug.  The issue was
that the recursive loop was stopping after two recursive depths and
throwing an error.  Instead, I created a bubble sort type algorithm to
sort the order of operations.
In addition, I removed the dependency on cfwheels when dumping data.